refactor arm-policy "externals" code
This is in preparation for integrating things with actix-web. The interpreter's "eval" function needs to be rewritten to return a future, which can then be run under actix. The current setup does not work because the evaluator tries to start its own tokio executor, which is blocked by actix. We get the following error:

   `panicked at 'Multiple executors at once: EnterError { reason: "attempted to run an executor while another executor is already running" }'`